# debt-service/internal/storage/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package storage

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type Debt struct {
	ID           uuid.UUID
	FirstName    string
	LastName     string
	PhoneNumber  string
	Jshshir      string
	Address      string
	BagID        string
	Price        float64
	PricePaid    float64
	Acquaintance sql.NullString
	Collateral   sql.NullString
	Deadline     time.Time
	Status       sql.NullInt32
	CreatedAt    sql.NullTime
	UpdatedAt    sql.NullTime
	DeletedAt    sql.NullInt64
}